Revert "新的内存管理模块 (#301)"
This reverts commit d8ad0a5e7724469abd5cc3cf271993538878033e.

+++ kernel/src/exception/mod.rs
@@ -19,15 +19,15 @@ pub trait InterruptArch: Send + Sync {
 
 #[derive(Debug, Clone, Copy)]
 pub struct IrqFlags {
-    flags: usize,
+    flags: u64,
 }
 
 impl IrqFlags {
-    pub fn new(flags: usize) -> Self {
+    pub fn new(flags: u64) -> Self {
         IrqFlags { flags }
     }
 
-    pub fn flags(&self) -> usize {
+    pub fn flags(&self) -> u64 {
         self.flags
     }
 }
